update mapping include custom serializers only if app installed move contact tests from base
